recipesPage/refactoring
